factor out room creator
