+++ src/lib/mapCreator.h
@@ -11,8 +11,6 @@
 #include "mapGeneration/BinarySpacePartitioner.h"
 #include "mapGeneration/Node.h"
 
-const int MAP_TALLNESS = 25;
-
 class MapCreator {
  public:
   int mapWidth, mapHeight;
@@ -27,13 +25,4 @@ class MapCreator {
   void start();
   void createMap();
   void createFloorModel(vec2 origin, int width, int height, TextureType textureType = OFFICE_FLOOR);
-};
-
-class RoomCreator {
- public:
-  RoomCreator(){};  // default constructor
-  void createRandPillarRoom(NodePtr room, int pillarSize = 10, int roomTallness = MAP_TALLNESS);
-
- private:
-  bool isOccupied(std::vector<std::vector<TileType>>& grid, int x, int y);
 };
